+++ .github/workflows/test.yml
@@ -54,5 +54,23 @@ jobs:
           echo "tests failed"
           exit 1
 
+  test-codegen:
+    name: Generated payloads build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          submodules: recursive
+
+      - uses: aave-dao/github-workflows/.github/actions/setup-node@main
+
+      - name: Run Foundry setup
+        uses: aave-dao/github-workflows/.github/actions/foundry-setup@main
+
+      # the snapshot specs only diff strings, so a payload that does not compile passes
+      # them; this builds the generator fixtures with forge
+      - name: Build generated payloads
+        run: pnpm test:codegen
+
   test-js:
     uses: aave-dao/github-workflows/.github/workflows/test-node.yml@main
