+++ generator/features/v4/access/accessManagerTargetFunctionRoleUpdate.ts
@@ -26,16 +26,22 @@ function rolesWiredHelper(market: MarketIdentifier): string {
   }`;
 }
 
+/// Name of the payload constant holding a raw-address target. Derived from the address
+/// rather than the config position, so the same target named by two features yields one
+/// constant instead of two colliding ones.
+function targetConstantName(c: V4TargetFunctionRoleUpdate): string {
+  return `ROLE_${sanitizeIdentifier(c.roleId)}_TARGET_${accessorIdentifier(c.target)}`;
+}
+
 /// Codegen for a target-function-role update's `target`: a raw address is emitted as
 /// a named constant; a codegen expr (accessor or payload constant) is used directly.
 function targetCodegen(
   market: MarketIdentifier,
   constants: string[],
   c: V4TargetFunctionRoleUpdate,
-  ix: number,
 ): string {
   if (c.target.startsWith('0x')) {
-    const targetName = `ROLE_${sanitizeIdentifier(c.roleId)}_TARGET_${ix}`;
+    const targetName = targetConstantName(c);
     constants.push(buildAddressConstant(market, targetName, c.target as `0x${string}`));
     return targetName;
   }
@@ -43,14 +49,13 @@ function targetCodegen(
 }
 
 /// Solidity-identifier-safe key for a target, used in generated test names.
-function targetIdentifier(c: V4TargetFunctionRoleUpdate, ix: number): string {
-  if (c.target.startsWith('0x')) return `ROLE_${sanitizeIdentifier(c.roleId)}_TARGET_${ix}`;
+function targetIdentifier(c: V4TargetFunctionRoleUpdate): string {
+  if (c.target.startsWith('0x')) return targetConstantName(c);
   return accessorIdentifier(c.target);
 }
 
-function targetTestCodegen(c: V4TargetFunctionRoleUpdate, ix: number): string {
-  if (c.target.startsWith('0x'))
-    return `proposal.ROLE_${sanitizeIdentifier(c.roleId)}_TARGET_${ix}()`;
+function targetTestCodegen(c: V4TargetFunctionRoleUpdate): string {
+  if (c.target.startsWith('0x')) return `proposal.${targetConstantName(c)}()`;
   if (c.target.includes('.')) return `address(${c.target})`;
   return `proposal.${c.target}()`;
 }
@@ -82,8 +87,8 @@ export const accessManagerTargetFunctionRoleUpdate: FeatureModule<V4TargetFuncti
   },
   build({market, cfg}) {
     const constants: string[] = [];
-    const entries = cfg.map((c, ix) => {
-      const targetExpr = targetCodegen(market, constants, c, ix);
+    const entries = cfg.map((c) => {
+      const targetExpr = targetCodegen(market, constants, c);
       if (c.selectorsExpr) {
         return `items[__INDEX__] = IConfigEngine.TargetFunctionRoleUpdate({
           authority: address(${market}.ACCESS_MANAGER),
@@ -103,40 +108,40 @@ export const accessManagerTargetFunctionRoleUpdate: FeatureModule<V4TargetFuncti
         });
       }`;
     });
-    const inputAsserts = cfg.flatMap((c, ix) => {
+    const inputAsserts = cfg.map((c) => {
       const lines = [
-        `assertEq(items[${ix}].authority, address(${market}.ACCESS_MANAGER), 'authority');`,
-        `assertEq(items[${ix}].target, ${targetTestCodegen(c, ix)}, 'target');`,
-        `assertEq(uint256(items[${ix}].roleId), uint256(${c.roleId}), 'roleId');`,
+        `assertEq(items[__INDEX__].authority, address(${market}.ACCESS_MANAGER), 'authority');`,
+        `assertEq(items[__INDEX__].target, ${targetTestCodegen(c)}, 'target');`,
+        `assertEq(uint256(items[__INDEX__].roleId), uint256(${c.roleId}), 'roleId');`,
       ];
       if (c.selectorAsserts) {
         lines.push(
-          `assertEq(items[${ix}].selectors.length, ${c.selectorAsserts.length}, 'selectors length');`,
+          `assertEq(items[__INDEX__].selectors.length, ${c.selectorAsserts.length}, 'selectors length');`,
         );
         c.selectorAsserts.forEach((sel, jx) =>
           lines.push(
-            `assertEq(uint32(items[${ix}].selectors[${jx}]), uint32(${sel}), 'selector ${jx}');`,
+            `assertEq(uint32(items[__INDEX__].selectors[${jx}]), uint32(${sel}), 'selector ${jx}');`,
           ),
         );
       }
-      return lines;
+      return lines.join('\n        ');
     });
-    const inputTest = `function test_accessManagerTargetFunctionRoleUpdatesInput() public view {
-        IConfigEngine.TargetFunctionRoleUpdate[] memory items = proposal.accessManagerTargetFunctionRoleUpdates();
-        assertEq(items.length, ${cfg.length}, 'length');
-        ${inputAsserts.join('\n        ')}
-      }`;
-    // One test per target: every role wired on the same contract is asserted together.
-    const byTarget = new Map<string, number[]>();
-    cfg.forEach((c, ix) => byTarget.set(c.target, [...(byTarget.get(c.target) ?? []), ix]));
-    const testFns = [...byTarget].map(([, ixs]) => {
-      const reference = cfg[ixs[0]].referenceTarget;
-      const referenceExpr = reference ? wrapAddress(reference) : 'address(0)';
-      const asserts = ixs.map((ix) => `_assertRolesWired(items[${ix}], ${referenceExpr});`);
-      return `function test_rolesWired_${targetIdentifier(cfg[ixs[0]], ixs[0])}() public {
+    // One test per target, matching on the target rather than on a config position:
+    // the getter also carries the wiring of every other feature in the payload.
+    const byTarget = new Map<string, V4TargetFunctionRoleUpdate>();
+    cfg.forEach((c) => byTarget.set(c.target, byTarget.get(c.target) ?? c));
+    const testFns = [...byTarget.values()].map((c) => {
+      const referenceExpr = c.referenceTarget ? wrapAddress(c.referenceTarget) : 'address(0)';
+      return `function test_rolesWired_${targetIdentifier(c)}() public {
         IConfigEngine.TargetFunctionRoleUpdate[] memory items = proposal.accessManagerTargetFunctionRoleUpdates();
         GovV3Helpers.executePayload(vm, address(proposal));
-        ${asserts.join('\n        ')}
+        uint256 matched;
+        for (uint256 i; i < items.length; ++i) {
+          if (items[i].target != ${targetTestCodegen(c)}) continue;
+          _assertRolesWired(items[i], ${referenceExpr});
+          ++matched;
+        }
+        assertGt(matched, 0, 'no wiring for target');
       }`;
     });
     const response: CodeArtifact = {
@@ -146,10 +151,11 @@ export const accessManagerTargetFunctionRoleUpdate: FeatureModule<V4TargetFuncti
           accessManagerTargetFunctionRoleUpdates: {
             returnType: 'IConfigEngine.TargetFunctionRoleUpdate',
             entries,
+            inputAsserts,
           },
         },
       },
-      test: {fn: [inputTest, ...testFns], helpers: [rolesWiredHelper(market)]},
+      test: {fn: testFns, helpers: [rolesWiredHelper(market)]},
     };
     return response;
   },
